Customized window

+++ src/constants.rs
@@ -1,8 +1,8 @@
 //! Constants used through-out the engine
 
 /// Default game window width
-pub const WINDOW_DEFAULT_WIDTH: f64 = 960.;
+pub const WINDOW_DEFAULT_WIDTH: f32 = 960.;
 /// Default game window height
-pub const WINDOW_DEFAULT_HEIGHT: f64 = 540.;
+pub const WINDOW_DEFAULT_HEIGHT: f32 = 540.;
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
